src/requirements.txt
connexion
docker
pyyaml
Jinja2
